refactoring decoders
