+++ TMDB/Source/Networking/Decoders/SnakeCaseDecoder.swift
@@ -1,0 +1,18 @@
+//
+//  SnakeCaseDecoder.swift
+//  TMDB
+//
+//  Created by Ruan Reis on 18/05/20.
+//  Copyright © 2020 Ruan Reis. All rights reserved.
+//
+
+import Foundation
+
+class SnakeCaseDecoder<T: Decodable>: UseDefaultDecoder<T> {
+    
+    override func decode(from data: Data) -> T? {
+        let decoder = JSONDecoder()
+        decoder.keyDecodingStrategy = .convertFromSnakeCase
+        return try? decoder.decode(expectation.self, from: data)
+    }
+}
